moved language selector to its own widget, adjusted styling of language selector
